requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.7.0
Pillow>=9.0.0
albumentations>=1.3.0
facenet-pytorch>=2.5.0
timm>=0.9.0
tensorboard>=2.12.0
matplotlib>=3.6.0
seaborn>=0.11.0
scikit-learn>=1.2.0
tqdm>=4.64.0
wandb>=0.15.0
pyyaml>=6.0
einops>=0.6.0
kornia>=0.7.0
